Move `tokenizer.py` to the folder of lmdeploy (#543)
* move tokenizer

* remove Tokenizer in init

* update deploy.py
